% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{-----------------------------------------------------------------------------
Data cleaning
-----------------------------------------------------------------------------
Clean earthquake data
This function cleans the earthquake data retrieved from NOAA database. It
uses the archived data including the records dated before 1995.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{data.frame -- the earthquake data base by NOAA (USA)}
}
\value{
data.frame
}
\description{
-----------------------------------------------------------------------------
Data cleaning
-----------------------------------------------------------------------------
Clean earthquake data
This function cleans the earthquake data retrieved from NOAA database. It
uses the archived data including the records dated before 1995.
}
\examples{
\dontrun{eq_clean_data(data)}
}
