% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{geom_timeline}
\alias{geom_timeline}
\title{-----------------------------------------------------------------------------
Visualization
-----------------------------------------------------------------------------
Earthquakes over Time
This function draws a point for each earthquake from the min to max date in
data. The size of points indicates the magnitude of earthquake and the color
indicates the number of deaths in each occurrence.}
\usage{
geom_timeline(data)
}
\arguments{
\item{data}{data.frame -- the earthquake data base by NOAA (USA)}
}
\value{
ggplot
}
\description{
-----------------------------------------------------------------------------
Visualization
-----------------------------------------------------------------------------
Earthquakes over Time
This function draws a point for each earthquake from the min to max date in
data. The size of points indicates the magnitude of earthquake and the color
indicates the number of deaths in each occurrence.
}
\examples{
\dontrun{
data <- eq_location_clean(eq_clean_data(data))
geom_timeline(data)}
}
